backend/requirements.txt
fastapi
uvicorn
openai
pydantic
motor
python-dotenv
elevenlabs
uagents
httpx
# Google APIs (Gmail + Calendar)
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
# Gemini (classifier)
google-genai
